Moved images to src
