+++ src/components/widget/d2l-teacher-course-creation-welcome.js
@@ -1,5 +1,5 @@
 import '@brightspace-ui/core/components/button/button.js';
-import '../../../images/create-course-illustration';
+import '../../images/create-course-illustration';
 import { css, html, LitElement } from 'lit-element/lit-element';
 import { BaseMixin } from '../../mixins/base-mixin';
 import { heading2Styles } from '@brightspace-ui/core/components/typography/styles.js';
